+++ src/main/java/com/cleanroommc/common/crafting/SmeltingRecipe.java
@@ -1,4 +1,4 @@
-package net.minecraftforge.common.crafting;
+package com.cleanroommc.common.crafting;
 
 import net.minecraft.item.ItemStack;
